Renaming VCR::WebSocket -> WebSocketVCR
