+++ lib/simple_websocket_vcr/configuration.rb
@@ -1,21 +1,19 @@
-module VCR
-  module WebSocket
-    class Configuration
-      attr_accessor :cassette_library_dir, :hook_uris, :json_cassettes
+module WebSocketVCR
+  class Configuration
+    attr_accessor :cassette_library_dir, :hook_uris, :json_cassettes
 
-      def initialize
-        reset_defaults!
-      end
+    def initialize
+      reset_defaults!
+    end
 
-      def reset_defaults!
-        @cassette_library_dir = 'spec/fixtures/vcr_cassettes'
-        @hook_uris = []
-        @json_cassettes = false
-      end
+    def reset_defaults!
+      @cassette_library_dir = 'spec/fixtures/vcr_cassettes'
+      @hook_uris = []
+      @json_cassettes = false
+    end
 
-      def method_missing(method_name, *_args, &_block)
-        puts 'unknown method: ' + method_name.to_s
-      end
+    def method_missing(method_name, *_args, &_block)
+      puts 'unknown method: ' + method_name.to_s
     end
   end
 end
