+++ lib/simple_websocket_vcr/errors.rb
@@ -1,11 +1,9 @@
-module VCR
-  module WebSocket
-    module Errors
-      class VCRError < StandardError; end
-      class OperationMismatchError < VCRError; end
-      class DataMismatchError < VCRError; end
-      class NoCassetteError < VCRError; end
-      class NoMoreSessionsError < VCRError; end
-    end
+module WebSocketVCR
+  module Errors
+    class VCRError < StandardError; end
+    class OperationMismatchError < VCRError; end
+    class DataMismatchError < VCRError; end
+    class NoCassetteError < VCRError; end
+    class NoMoreSessionsError < VCRError; end
   end
 end
