+++ lib/simple_websocket_vcr/version.rb
@@ -1,5 +1,3 @@
-module VCR
-  module WebSocket
-    VERSION = '0.0.5'
-  end
+module WebSocketVCR
+  VERSION = '0.0.5'
 end
